+++ src/main/java/roomescape/reservation/domain/ReservationTime.java
@@ -1,12 +1,12 @@
-package roomescape.reservation.dto;
+package roomescape.reservation.domain;
 
 import java.time.LocalTime;
 
-public class ReservationTimeRequest {
+public class ReservationTime {
     private final Long id;
     private final LocalTime startAt;
 
-    public ReservationTimeRequest(final Long id, final LocalTime startAt) {
+    public ReservationTime(final Long id, final LocalTime startAt) {
         this.id = id;
         this.startAt = startAt;
     }
