+++ src/main/java/roomescape/reservation/presentation/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.controller;
+package roomescape.reservation.presentation;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.reservation.dto.ReservationRequest;
-import roomescape.reservation.dto.ReservationResponse;
-import roomescape.reservation.service.ReservationService;
+import roomescape.reservation.presentation.dto.ReservationRequest;
+import roomescape.reservation.presentation.dto.ReservationResponse;
+import roomescape.reservation.application.ReservationService;
 
 @RestController
 @RequestMapping("/reservations")
